+++ src/app/lib/time.ts
@@ -1,12 +1,28 @@
 import * as moment from 'moment'
 
-export function formatDuration(unixTime: number): string {
+export function formatDuration(unixTime: number, short: boolean = false): string {
+  return short ? formatDurationShort(unixTime) : formatDurationLong(unixTime)
+}
+
+export function formatDurationLong(unixTime: number): string {
   return moment.unix(unixTime).utc().format('HH:mm:ss.SSS')
+}
+
+export function formatDurationCombined(unixTime: number, seconds: boolean = false): string {
+  return seconds ? Math.round(unixTime) + ' | ' + moment.unix(unixTime).utc().format('m:ss') : formatDurationLong(unixTime)
+}
+
+export function formatDurationShort(unixTime: number): string {
+  return Math.round(unixTime) + ''
 }
 
 const parseDurationRegex = /^(?:(?:([0-9]*):)|(?:([0-9]*):([0-9]*):))?([0-9]*)(?:\.([0-9]*))?$/
 
-export function parseDuration(durationStr: string): number {
+export function parseDuration(durationStr: string, short: boolean = false): number {
+  return short ? parseDurationShort(durationStr) : parseDurationLong(durationStr)
+}
+
+export function parseDurationLong(durationStr: string): number {
   const result = parseDurationRegex.exec(durationStr)
   if(result === null) {
     return 0
@@ -18,4 +34,9 @@ export function parseDuration(durationStr: string): number {
     const sFract = result[5] ? parseFloat('.'+result[5]) : 0
     return h*3600 + m1*60 + m2*60 + s + sFract
   }
+}
+
+export function parseDurationShort(durationStr: string): number {
+  const items = durationStr.split(' | ')
+  return parseFloat(items[0])
 }